improved impelementation of the marching cubes algorithm add new chapter to the theses

+++ MarchingCubesVisualisation/fileHandler.h
@@ -10,7 +10,7 @@
 #define swap_long(num) (((num >> 24) & 0xff) | ((num << 8) & 0xff0000) | ((num >> 8) & 0xff00) | ((num << 24) & 0xff000000)) 
 
 typedef enum fileType{img, hdr};
-typedef std::vector<std::vector<std::vector<double>>> d3Buffer;
+typedef std::vector<std::vector<std::vector<float>>> d3Buffer;
 
 class fileHandler{
 public:
